+++ aiworkspace-20210204/alibabacloud_aiworkspace20210204/client.py
@@ -63,14 +63,8 @@ class Client(OpenApiClient):
             'eu-west-1-oxs': 'aiworkspace.aliyuncs.com',
             'me-east-1': 'aiworkspace.aliyuncs.com',
             'rus-west-1-pop': 'aiworkspace.aliyuncs.com',
-            'us-west-1': 'aiworkspace.us-west-1.aliyuncs.com',
-            'us-southeast-1': 'aiworkspace.us-southeast-1.aliyuncs.com',
-            'us-east-1': 'aiworkspace.us-east-1.aliyuncs.com',
-            'na-south-1': 'aiworkspace.na-south-1.aliyuncs.com',
-            'eu-central-1': 'aiworkspace.eu-central-1.aliyuncs.com',
             'cn-wulanchabu': 'aiworkspace.cn-wulanchabu.aliyuncs.com',
             'cn-shenzhen': 'aiworkspace.cn-shenzhen.aliyuncs.com',
-            'cn-shanghai-finance-1': 'aiworkspace.cn-shanghai-finance-1.aliyuncs.com',
             'cn-shanghai': 'aiworkspace.cn-shanghai.aliyuncs.com',
             'cn-hongkong': 'aiworkspace.cn-hongkong.aliyuncs.com',
             'cn-heyuan': 'aiworkspace.cn-heyuan.aliyuncs.com',
@@ -81,9 +75,15 @@ class Client(OpenApiClient):
             'ap-southeast-7': 'aiworkspace.ap-southeast-7.aliyuncs.com',
             'ap-southeast-3': 'aiworkspace.ap-southeast-3.aliyuncs.com',
             'ap-southeast-1': 'aiworkspace.ap-southeast-1.aliyuncs.com',
-            'ap-south-1': 'aiworkspace.ap-south-1.aliyuncs.com',
             'ap-northeast-2': 'aiworkspace.ap-northeast-2.aliyuncs.com',
-            'ap-northeast-1': 'aiworkspace.ap-northeast-1.aliyuncs.com'
+            'ap-northeast-1': 'aiworkspace.ap-northeast-1.aliyuncs.com',
+            'eu-central-1': 'aiworkspace.eu-central-1.aliyuncs.com',
+            'na-south-1': 'aiworkspace.na-south-1.aliyuncs.com',
+            'us-east-1': 'aiworkspace.us-east-1.aliyuncs.com',
+            'us-southeast-1': 'aiworkspace.us-southeast-1.aliyuncs.com',
+            'us-west-1': 'aiworkspace.us-west-1.aliyuncs.com',
+            'ap-south-1': 'aiworkspace.ap-south-1.aliyuncs.com',
+            'cn-shanghai-finance-1': 'aiworkspace.cn-shanghai-finance-1.aliyuncs.com'
         }
         self.check_config(config)
         self._endpoint = self.get_endpoint('aiworkspace', self._region_id, self._endpoint_rule, self._network, self._suffix, self._endpoint_map, self._endpoint)
@@ -210,6 +210,8 @@ class Client(OpenApiClient):
             body['SourceId'] = request.source_id
         if not DaraCore.is_null(request.source_type):
             body['SourceType'] = request.source_type
+        if not DaraCore.is_null(request.user_id):
+            body['UserId'] = request.user_id
         if not DaraCore.is_null(request.workspace_id):
             body['WorkspaceId'] = request.workspace_id
         req = open_api_util_models.OpenApiRequest(
@@ -258,6 +260,8 @@ class Client(OpenApiClient):
             body['SourceId'] = request.source_id
         if not DaraCore.is_null(request.source_type):
             body['SourceType'] = request.source_type
+        if not DaraCore.is_null(request.user_id):
+            body['UserId'] = request.user_id
         if not DaraCore.is_null(request.workspace_id):
             body['WorkspaceId'] = request.workspace_id
         req = open_api_util_models.OpenApiRequest(
@@ -2691,6 +2695,8 @@ class Client(OpenApiClient):
             body['RoleName'] = request.role_name
         if not DaraCore.is_null(request.role_type):
             body['RoleType'] = request.role_type
+        if not DaraCore.is_null(request.user_id):
+            body['UserId'] = request.user_id
         req = open_api_util_models.OpenApiRequest(
             headers = headers,
             body = Utils.parse_to_map(body)
@@ -2726,6 +2732,8 @@ class Client(OpenApiClient):
             body['RoleName'] = request.role_name
         if not DaraCore.is_null(request.role_type):
             body['RoleType'] = request.role_type
+        if not DaraCore.is_null(request.user_id):
+            body['UserId'] = request.user_id
         req = open_api_util_models.OpenApiRequest(
             headers = headers,
             body = Utils.parse_to_map(body)