+++ aiworkspace-20210204/alibabacloud_aiworkspace20210204/models/_create_run_response_body.py
@@ -10,7 +10,7 @@ class CreateRunResponseBody(DaraModel):
         run_id: str = None,
         request_id: str = None,
     ):
-        # The ID of the run.
+        # The run ID.
         self.run_id = run_id
         # Id of the request
         self.request_id = request_id